Refactor managercli
Moved each command class and methods to individual modules for more
effective maintenance
