+++ src/subscription_manager/scripts/subscription_manager.py
@@ -69,7 +69,7 @@ try:
     inj.require(inj.DBUS_IFACE)
 
     from subscription_manager import managercli
-    from subscription_manager.managercli import handle_exception
+    from subscription_manager.cli_command.cli import handle_exception
 
     from subscription_manager import ga_loader
     ga_loader.init_ga()
